Move fetch call for FAQ's from App to Home component
